% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaTables.R
\docType{package}
\name{metaTables}
\alias{metaTables}
\alias{metaTables-package}
\title{metaTables: A package for streamlining the computation of meta-regressions and sub-groups analysis on a series of variables.}
\description{
Run meta-regressions and sub-groups analysis on lists of variables and produce tables for publication.
}
\section{functions}{

The three main functions include:

\link[metaTables]{regTable}: This function allows you to perform meta-regressions on a list of variables and output the results in a table

\link[metaTables]{subGroups}: This function allows you to perform sub-groups on all levels of a variable and output the results in a table

\link[metaTables]{propTable}: This function allows you to create a proportions table with frequencies and percentages broken down by variables
}

