% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/propTable.R
\name{propTable}
\alias{propTable}
\title{propTable}
\usage{
propTable(data = NULL, variables = NULL, study.id = NULL,
  labels = NULL)
}
\arguments{
\item{data}{The dataset}

\item{variables}{The variables that you  would like a proportion table of}

\item{study.id}{The study identifier}

\item{labels}{The labels that can be used in place of the variable names}
}
\value{
A table of proportions and frequencies
}
\description{
This function allows you to create a proportions table with frequencies and percentages broken down by variables
}
