% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regTable.R
\name{regTable}
\alias{regTable}
\title{regTable}
\usage{
regTable(append.table = FALSE, data = NULL, moderators = NULL,
  es = NULL, vi = NULL, study.id = NULL, labels = NULL,
  modelweights = "CORR", ...)
}
\arguments{
\item{append.table}{Determines whether to append to an exisitng table. By default it is set to FALSE and will create a new table or overide the existing object. To append specify the existing table object}

\item{moderators}{The variables used for the meta-regressions. Each moderator will be used for its own meta-regresssion.}

\item{es}{The effect size variable}

\item{study.id}{The study identifier}

\item{labels}{The labels that will be added to the top of each table. The variable name is used if unspecified}

\item{modelweights}{Either "CORR" or "HEIR" implmentation of Robust Variance Estimation. "CORR" is set as default}

\item{yi}{The variance in the effect size}
}
\value{
A table of a series of meta-regressions
}
\description{
This function allows you to perform meta-regressions on a list of variables and output the results in a table
}
\details{
#' @param data The dataset
}
