% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regTitle.R
\name{regTitle}
\alias{regTitle}
\title{regTitle}
\usage{
regTitle(label, headings = NULL, append.table = FALSE)
}
\arguments{
\item{label}{The label that will be added to the table}

\item{headings}{Set to NULL by default. Can be used to specify custom headings. The default is headings are "Var","b","SE","ciL","ciU", "p"}

\item{append.table}{Determines whether to append to an exisitng table. By default it is set to false and will create a new table or overide the existing object. To append specify the existing table object}
}
\value{
A table with a single row with the label or a label appended to existing table
}
\description{
This function allows you to add a single line to a meta-regression table with a label for identifying.
Labels are added automatically by the regTable function, but these may be of further use for deliniating sections of the table
}
