% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subGroups.R
\name{subGroups}
\alias{subGroups}
\title{subGroups}
\usage{
subGroups(append.table = FALSE, data = NULL, grouping.factor = NULL,
  es = NULL, vi = NULL, study.id = NULL, label = NULL,
  modelweights = "CORR", type = NULL, headings = NULL, ...)
}
\arguments{
\item{append.table}{Determines whether to append to an exisitng table. By default it is set to FALSE and will create a new table or overide the existing object. To append specify the existing table object}

\item{grouping.factor}{The variables used for the sub-groups. All levels of each grouping factor will be run as a sub-group}

\item{es}{The effect size variable}

\item{study.id}{The study identifier}

\item{label}{The labels that will be added to the top of each table section. The grouping.factor variable name is used if unspecified}

\item{modelweights}{Either "CORR" or "HEIR" implmentation of Robust Variance Estimation. "CORR" is set as default}

\item{type}{Either "cor" or "MD". Determines the headings used if not user specified.}

\item{headings}{Set to NULL by default. Can be used to specify custom headings. If unspecified then headings are determined by the type argument}

\item{yi}{The variance in the effect size}
}
\value{
A table of a series of sub-groups analyses
}
\description{
This function allows you to perform sub-groups on all levels of a variable and output the results in a table
}
\details{
#' @param data The dataset
}
