% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subTitle.R
\name{subTitle}
\alias{subTitle}
\title{subTitle}
\usage{
subTitle(label, type = NULL, headings = NULL, append.table = FALSE)
}
\arguments{
\item{label}{The label that will be added to the table}

\item{type}{Either "cor" or "MD". Determines the headings used if not user specified.}

\item{headings}{Set to NULL by default. Can be used to specify custom headings. If unspecified then headings are determined by the type argument}

\item{append.table}{Determines whether to append to an exisitng table. By default it is set to false and will create a new table or overide the existing object. To append specify the existing table object}
}
\value{
A table with a single row with the label or a label appended to existing table
}
\description{
This function allows you to add a single line to a sub-group analysis table  with a label for identifying.
Labels are added automatically by the subGroup function, but these may be of further use for deliniating sections of the table
}
